Change class name - input list
